requirements.txt
Flask==3.0.3
Pillow==10.1.0